# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OTO-RAPORT API - Ministry Compliance & Developer Management",
    "version": "1.0.0",
    "description": "REST API for OTO-RAPORT SaaS platform. Provides endpoints for Ministry Schema 1.13 compliance validation, developer profile management, and property data validation.\n\n**Authentication**: All endpoints require Supabase Auth session (JWT token in Authorization header).\n\n**Ministry Compliance**: Based on Polish law 'ustawa z dnia 21 maja 2025 r. o jawności cen mieszkań'.",
    "contact": {
      "name": "OTO-RAPORT Support",
      "email": "support@oto-raport.pl"
    }
  },
  "servers": [
    {
      "url": "https://otoraport-v2.vercel.app/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/developers/{id}/completion-status": {
      "get": {
        "tags": ["Developers"],
        "summary": "Get developer profile completion status",
        "description": "Returns detailed completion status including section breakdown, missing fields, and actionable next steps.",
        "operationId": "getDeveloperCompletionStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Developer UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completion status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionStatusResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "developerId": "123e4567-e89b-12d3-a456-426614174000",
                    "companyName": "Deweloper XYZ Sp. z o.o.",
                    "overallCompletion": 72,
                    "sectionCompletion": {
                      "basicInfo": {
                        "complete": true,
                        "percentage": 100
                      },
                      "headquarters": {
                        "complete": false,
                        "percentage": 75
                      },
                      "salesOffice": {
                        "complete": false,
                        "percentage": 50
                      },
                      "contact": {
                        "complete": true,
                        "percentage": 100
                      }
                    },
                    "missingCriticalFields": [],
                    "missingRecommendedFields": [
                      "REGON",
                      "Ulica siedziby",
                      "Miasto biura sprzedaży"
                    ],
                    "nextSteps": [
                      "📋 Uzupełnij podstawowe informacje o firmie (75% gotowe)",
                      "🏪 Dodaj adres biura sprzedaży (50% gotowe)"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "SupabaseAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompletionStatusResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "required": ["developerId", "companyName", "overallCompletion", "sectionCompletion", "missingCriticalFields", "missingRecommendedFields", "nextSteps"],
            "properties": {
              "developerId": {
                "type": "string",
                "format": "uuid"
              },
              "companyName": {
                "type": "string",
                "example": "Deweloper XYZ Sp. z o.o."
              },
              "overallCompletion": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100,
                "example": 72
              },
              "sectionCompletion": {
                "type": "object",
                "properties": {
                  "basicInfo": {
                    "$ref": "#/components/schemas/SectionCompletionStatus"
                  },
                  "headquarters": {
                    "$ref": "#/components/schemas/SectionCompletionStatus"
                  },
                  "salesOffice": {
                    "$ref": "#/components/schemas/SectionCompletionStatus"
                  },
                  "contact": {
                    "$ref": "#/components/schemas/SectionCompletionStatus"
                  }
                }
              },
              "missingCriticalFields": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": []
              },
              "missingRecommendedFields": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["REGON", "Ulica siedziby"]
              },
              "nextSteps": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "📋 Uzupełnij podstawowe informacje o firmie (75% gotowe)",
                  "🏪 Dodaj adres biura sprzedaży (50% gotowe)"
                ]
              }
            }
          }
        }
      },
      "SectionCompletionStatus": {
        "type": "object",
        "required": ["complete", "percentage"],
        "properties": {
          "complete": {
            "type": "boolean",
            "example": false
          },
          "percentage": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "example": 75
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Internal server error"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - validation errors",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Validation failed",
              "details": [
                {
                  "field": "updates.nip",
                  "message": "NIP musi składać się z 10 cyfr"
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Forbidden - not your developer profile"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Developer not found"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "SupabaseAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token obtained from authentication. Include in Authorization header: `Bearer {token}`"
      }
    }
  }
}